refactor: split to components
