+++ src/App.css
@@ -33,7 +33,6 @@ main,
 main {
     display: flex;
     flex-direction: column;
-    border: 2px solid red;
 }
 
 table,
